+++ types/vergeSense/cloudConnector/uplink.spec.js
@@ -38,6 +38,14 @@ describe("Verge Sense Uplink", () => {
       });
   });
 
+  let sensorReportSchema = null;
+  before((done) => {
+    loadSchema(`${__dirname}/sensor_report.schema.json`)
+      .then((parsedSchema) => {
+        sensorReportSchema = parsedSchema;
+        done();
+      });
+  });
 
   describe("consume()", () => {
     it("should decode the Verge Sense space report payload && init state", () => {
@@ -209,6 +217,42 @@ describe("Verge Sense Uplink", () => {
         assert.equal(value.data.peopleCount, 150);
 
         validateSchema(value.data, areaCountReportSchema, {
+          throwError: true,
+        });
+      });
+
+      consume(data);
+    });
+
+    it("should decode the Verge Sense sensor report", () => {
+      const data = {
+        data: {
+          "building_ref_id": "SYLE",
+          "event_type": "sensor_report",
+          "floor_ref_id": "23",
+          "person_count": 0,
+          "sensor_id": "3BE-4NZ",
+          "signs_of_life": false,
+          "space_ref_id": "23. C0.010",
+          "timestamp": "2026-04-08T16:44:13.000Z"
+        }
+      };
+
+      expectEmits((type, value) => {
+        assert.equal(type, "state");
+        assert.isNotNull(value);
+      });
+
+
+      expectEmits((type, value) => {
+        assert.equal(type, "sample");
+        assert.isNotNull(value);
+        assert.typeOf(value.data, "object");
+
+        assert.equal(value.topic, "sensor_report");
+        assert.equal(value.data.signsOfLife, false);
+
+        validateSchema(value.data, sensorReportSchema, {
           throwError: true,
         });
       });